4.21 done

+++ osa4/blog list/controllers/blogs.js
@@ -113,8 +113,24 @@ blogsRouter.delete('/:id', async (request, response, next) => {
     //     }
     // Test express-async-errors library!
     // Do this to other routes!
+
+    //4.21*
+    const blog = await Blog.findById(request.params.id)
+    if (!blog) {
+        return response.status(404).json({ error: 'blog not found' })
+    }
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    if (blog.user.toString() !== decodedToken.id.toString()) {
+        return response.status(401).json({ error: 'only the creator can delete blogs' })
+    }
+
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
+
+
 })
 
 //4.14
